Fix neighbour_alias DB Result logic for Device Port
We want the single device_ip row that matches the IP in remote_ip
on the device port.

Previously, the logic was to get all the device_ip rows for the
remote device connected to the device port. This doesn't match the
belongs_to relationship expressed by neighbour_alias or the
documentation.